requirements.txt
# IntellGrade Python Dependencies
# Install with: pip install -r requirements.txt

# Core data science libraries
numpy>=1.21.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0

# Database
mysql-connector-python>=8.0.0
pymysql>=1.0.0

# Web framework for API
flask>=2.0.0
flask-cors>=3.0.0

# Security
bcrypt>=3.2.0
python-dotenv>=0.19.0

# Optional: For enhanced AI features
# nltk>=3.6.0
# matplotlib>=3.4.0
# seaborn>=0.11.0

# Development and testing (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0 